Updated Entity and some minor chages in other classes

+++ src/main/java/com/rewardly/emp/controller/EmployeeController.java
@@ -22,7 +22,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class EmployeeController {
 
-	
 	private final EmployeeService employeeService;
 
 	@PostMapping("/create")
@@ -30,35 +29,35 @@ public class EmployeeController {
 
 		return ResponseEntity.ok(employeeService.createEmployee(employee));
 	}
-	
-	
-	
+
 	@GetMapping("/hello")
 	public String hello() {
 		return "Hello from Employee Service";
 	}
-	
+
 	@GetMapping("/{id}")
-	public Employee getEmployee(@PathVariable int id) {
+	public Employee getEmployee(@PathVariable Long id) {
 		Employee employee = employeeService.getEmployee(id);
 		return employee;
 	}
+
 	@GetMapping
-	public List<Employee> getAllEmployees(){
+	public List<Employee> getAllEmployees() {
 		List<Employee> allEmployees = employeeService.getAllEmployees();
 		return allEmployees;
-		
+
 	}
+
 	@PutMapping("/{id}")
-	public Employee updateEmployee(@PathVariable int id,@RequestBody Employee employee) {
-		Employee updateEmployee = employeeService.updateEmployee(id,employee);
+	public Employee updateEmployee(@PathVariable Long id, @RequestBody Employee employee) {
+		Employee updateEmployee = employeeService.updateEmployee(id, employee);
 		return updateEmployee;
 	}
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<String> deleteEmployee(@PathVariable int id){
+	public ResponseEntity<String> deleteEmployee(@PathVariable Long id) {
 		String message = employeeService.deleteEmployee(id);
 		return ResponseEntity.ok(message);
 	}
-	
 
 }
